@@ -1,110 +1,68 @@
 {
   "name": "cropperjs",
-  "description": "JavaScript image cropper.",
-  "version": "2.0.0-alpha.2",
-  "main": "dist/cropper.common.js",
-  "module": "dist/cropper.esm.js",
-  "browser": "dist/cropper.js",
-  "types": "types/index.d.ts",
-  "style": "dist/cropper.css",
-  "files": [
-    "src",
-    "dist",
-    "types"
+  "version": "2.0.0-beta",
+  "private": true,
+  "workspaces": [
+    "packages/*"
   ],
   "scripts": {
-    "build": "npm run build:css && npm run build:js",
-    "build:css": "postcss src/index.css -o dist/cropper.css --no-map",
-    "build:js": "rollup -c",
-    "clean": "del-cli dist",
-    "compress": "npm run compress:css && npm run compress:js",
-    "compress:css": "postcss dist/cropper.css -u cssnano -o dist/cropper.min.css --no-map",
-    "compress:js": "uglifyjs dist/cropper.js -o dist/cropper.min.js -c -m --comments /^!/",
-    "copy": "cpy dist/cropper.css docs/css",
+    "bootstrap": "lerna bootstrap",
+    "build": "npm run bootstrap && vuepress build docs",
+    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
     "lint": "npm run lint:js && npm run lint:css",
-    "lint:css": "stylelint **/*.{css,scss,html} --fix",
-    "lint:js": "eslint . --fix",
+    "lint:css": "stylelint **/*.{css,scss,vue} --fix",
+    "lint:js": "eslint . docs/.vuepress --fix",
     "prepare": "husky install",
-    "release": "npm run clean && npm run lint && npm run build && npm run compress && npm run copy && npm test",
-    "start": "npm-run-all --parallel watch:*",
-    "test": "karma start",
-    "test:coverage": "cat coverage/lcov.info | codecov",
-    "watch:css": "postcss src/index.css -o docs/css/cropper.css -m -w",
-    "watch:js": "rollup -c -m -w"
+    "release": "npm run bootstrap && lerna run release && npm run lint && npm test && npm run build && npm run changelog",
+    "start": "npm run bootstrap && vuepress dev docs",
+    "test": "jest"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/fengyuanchen/cropperjs.git"
-  },
-  "keywords": [
-    "image",
-    "crop",
-    "move",
-    "zoom",
-    "rotate",
-    "scale",
-    "cropper",
-    "cropper.js",
-    "cropping",
-    "processing",
-    "html",
-    "css",
-    "javascript",
-    "front-end",
-    "web"
-  ],
-  "author": {
-    "name": "Chen Fengyuan",
-    "url": "https://chenfengyuan.com/"
-  },
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/fengyuanchen/cropperjs/issues"
-  },
-  "homepage": "https://fengyuanchen.github.io/cropperjs",
   "devDependencies": {
-    "@babel/core": "^7.16.5",
-    "@babel/preset-env": "^7.16.5",
-    "@commitlint/cli": "^15.0.0",
-    "@commitlint/config-conventional": "^15.0.0",
-    "@rollup/plugin-babel": "^5.3.0",
-    "babel-plugin-istanbul": "^6.1.1",
-    "chai": "^4.3.4",
+    "@commitlint/cli": "^16.2.3",
+    "@commitlint/config-conventional": "^16.2.1",
+    "@microsoft/api-extractor": "^7.20.0",
+    "@rollup/plugin-commonjs": "^21.0.3",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-replace": "^4.0.0",
+    "@rollup/plugin-typescript": "^8.3.1",
+    "@types/estree": "^0.0.51",
+    "@types/jest": "^27.4.1",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "@vuepress/plugin-google-analytics": "^2.0.0-beta.37",
+    "@vuepress/plugin-pwa": "^2.0.0-beta.37",
+    "@vuepress/plugin-pwa-popup": "^2.0.0-beta.37",
+    "@vuepress/plugin-register-components": "^2.0.0-beta.37",
+    "@vuepress/plugin-search": "^2.0.0-beta.37",
     "change-case": "^4.1.2",
-    "codecov": "^3.8.2",
-    "cpy-cli": "^3.1.1",
+    "conventional-changelog-cli": "^2.2.2",
     "create-banner": "^2.0.0",
-    "cssnano": "^5.0.14",
+    "cssnano": "^5.1.7",
     "del-cli": "^4.0.1",
-    "eslint": "^8.5.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint": "^8.12.0",
+    "eslint-config-airbnb-typescript": "^16.2.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsdoc": "^38.1.6",
+    "eslint-plugin-vue": "^8.5.0",
     "husky": "^7.0.4",
-    "karma": "^6.3.9",
-    "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^2.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-rollup-preprocessor": "^7.0.7",
-    "lint-staged": "^12.1.4",
-    "mocha": "^9.1.3",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.5",
+    "jest": "^27.5.1",
+    "lerna": "^4.0.0",
+    "lint-staged": "^12.3.7",
+    "postcss": "^8.4.12",
     "postcss-cli": "^9.1.0",
-    "postcss-header": "^3.0.1",
     "postcss-html": "^1.3.0",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^7.1.0",
-    "postcss-url": "^10.1.3",
-    "puppeteer": "^13.0.1",
-    "rollup": "^2.62.0",
-    "stylelint": "^14.2.0",
-    "stylelint-config-standard-scss": "^3.0.0",
+    "rgb-hex": "^4.0.0",
+    "rollup": "^2.70.1",
+    "rollup-plugin-inline-postcss": "^3.0.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "stylelint": "^14.6.1",
+    "stylelint-config-recommended-scss": "^6.0.0",
+    "stylelint-config-recommended-vue": "^1.4.0",
     "stylelint-order": "^5.0.0",
-    "uglify-js": "^3.14.5"
-  },
-  "publishConfig": {
-    "tag": "next"
+    "ts-jest": "^27.1.4",
+    "ts-loader": "^9.2.8",
+    "tslib": "^2.3.1",
+    "typescript": "^4.6.3",
+    "vuepress": "^2.0.0-beta.37"
   }
 }
