@@ -1,30 +1,55 @@
 module.exports = {
-  root: true,
-  extends: 'airbnb-base',
   env: {
     browser: true,
+    node: true,
   },
+  extends: [
+    'airbnb-typescript/base',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:jsdoc/recommended',
+    'plugin:vue/vue3-recommended',
+  ],
+  parser: 'vue-eslint-parser',
+  parserOptions: {
+    parser: '@typescript-eslint/parser',
+    project: 'tsconfig.json',
+    sourceType: 'module',
+    extraFileExtensions: ['.vue'],
+  },
+  plugins: [
+    '@typescript-eslint',
+    'import',
+    'jsdoc',
+    'vue',
+  ],
   rules: {
+    '@typescript-eslint/no-explicit-any': 'off',
     'import/no-extraneous-dependencies': 'off',
-    'no-param-reassign': 'off',
-    'no-restricted-properties': 'off',
-    'valid-jsdoc': ['error', {
-      requireReturn: false,
+    'max-len': ['error', 100, 2, {
+      ignorePattern: '\\S+="[^"]*"',
+      ignoreComments: true,
+      ignoreUrls: true,
+      ignoreStrings: true,
+      ignoreTemplateLiterals: true,
+      ignoreRegExpLiterals: true,
     }],
+    'no-param-reassign': 'off',
   },
   overrides: [
     {
-      files: 'test/**/*.spec.js',
+      files: 'packages/**/tests/**/*.ts',
       env: {
-        mocha: true,
-      },
-      globals: {
-        Cropper: true,
-        expect: true,
+        jest: true,
       },
       rules: {
         'no-new': 'off',
-        'no-unused-expressions': 'off',
+      },
+    },
+    {
+      files: 'docs/.vuepress/**/*.vue',
+      rules: {
+        'import/no-unresolved': 'off',
+        'no-new': 'off',
       },
     },
   ],
