@@ -1,4 +1,4 @@
 module.exports = {
-  '*.js': 'eslint --fix',
-  '*.{css,scss,html}': 'stylelint --fix',
+  '*.{js,ts,vue}': 'eslint --fix',
+  '*.{css,scss,vue}': 'stylelint --fix',
 };
