@@ -1,15 +1,11 @@
 module.exports = {
-  extends: 'stylelint-config-recommended-scss',
+  extends: [
+    'stylelint-config-recommended-vue/scss',
+  ],
   plugins: [
     'stylelint-order',
   ],
   rules: {
     'order/properties-alphabetical-order': true,
   },
-  overrides: [
-    {
-      files: ['**/*.html'],
-      customSyntax: 'postcss-html',
-    },
-  ],
 };
